tls: move tls::ExData to tls/mod.rs
